style: Rename APIErrorResponse.ts to APIError.ts
This new name matches with the name of the actual type it exports.
